pyproject.toml
[project]
name = "astra"
version = "0.1.0"
description = "Agentic Skill & TRAce — A data factory for high-quality multi-turn tool-use conversation trajectories"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Astra Team" }
]
keywords = ["ai", "agent", "fine-tuning", "tool-use", "trajectories"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "hydra-core>=1.3",
    "rich>=13.0",
    "openai>=2.26.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/astra"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
